% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_api_key}
\alias{.check_api_key}
\title{Fix improper API keys}
\usage{
.check_api_key(api_key)
}
\arguments{
\item{api_key}{A \code{string} value containing a TERN API key for checking}
}
\value{
A \code{string} value with replacement of troublesome characters if
necessary.
}
\description{
Fix improper API keys
}
\keyword{internal}
