% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_date}
\alias{.check_date}
\title{Check User Input Dates for Validity}
\usage{
.check_date(x)
}
\arguments{
\item{x}{User entered date value}
}
\value{
Validated date string as a \code{POSIXct} object.
}
\description{
Check User Input Dates for Validity
}
\note{
This was taken from \CRANpkg{nasapower}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
.check_date("2024-01-01")
\dontshow{\}) # examplesIf}
}
\author{
Adam H. Sparks \email{adamhsparks@curtin.edu.au}
}
\keyword{internal}
