% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R, R/read_smips.R
\name{.check_not_example_api_key}
\alias{.check_not_example_api_key}
\title{Check that the user hasn't blindly copied the "your_api_key" string from the
examples}
\usage{
.check_not_example_api_key(.api_key)

.check_not_example_api_key(.api_key)
}
\description{
Check that the user hasn't blindly copied the "your_api_key" string from the
examples

Check that the user hasn't blindly copied the "your_api_key" string from the
examples
}
\keyword{internal}
\keyword{internal}
