% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_tern_data.R
\name{.normalise_vector_elements}
\alias{.normalise_vector_elements}
\title{Helper function: resolve the given \code{character} vector to its valid elements}
\usage{
.normalise_vector_elements(vec, valid_elements, info)
}
\arguments{
\item{vec}{A \code{character} vector (of user-supplied elements, e.g., dataset
names, variants), or \code{NULL}/\code{"all"}.}

\item{valid_elements}{A \code{character} vector containing the valid
elements against which the \code{given_vec} vector should be compared.}

\item{info}{A \code{character} string that describes the elements
in \code{given_vec} (so that errors/warnings are more informative).}
}
\value{
A \code{character} vector of valid elements (de-duplicated,
order preserved).
}
\description{
Helper function: resolve the given \code{character} vector to its valid elements
}
\keyword{internal}
