% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_tern_data.R
\name{.parse_date_range}
\alias{.parse_date_range}
\title{Normalise a date_range argument}
\usage{
.parse_date_range(date_range)
}
\arguments{
\item{date_range}{A \code{Date} vector or \code{character} vector containing the
date range supplied by the user.}
}
\value{
A \code{Date} vector.
}
\description{
If a \code{Date} vector or \code{character} vector of length 2 is given, parses
into a contiguous range of dates. If a vector of length greater than 2
is given, this function does nothing (besides the date conversion).
}
\keyword{internal}
