% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\docType{data}
\name{.phenology_metrics}
\alias{.phenology_metrics}
\title{Phenology metric registry}
\format{
A named \code{list} of \code{character} strings.
}
\usage{
.phenology_metrics
}
\description{
Named list mapping metric abbreviations to their subdirectory names on
the TERN data server.
}
\keyword{datasets}
\keyword{internal}
