% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\docType{data}
\name{.SLGA_META}
\alias{.SLGA_META}
\title{SLGA metadata: version, suffix, date for each soil attribute}
\format{
An object of class \code{list} of length 8.
}
\usage{
.SLGA_META
}
\description{
SLGA metadata: version, suffix, date for each soil attribute
}
\keyword{datasets}
