% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gtiff.R
\name{download_gtiff}
\alias{download_gtiff}
\title{Download a Cloud Optimised Geotiff (\acronym{COG}) file to your local
device.}
\usage{
download_gtiff(
  data = "smips",
  collection = "totalbucket",
  day,
  api_key = get_key(),
  file = getwd(),
  quiet = FALSE,
  cache = TRUE
)
}
\arguments{
\item{collection}{A character vector of the data collection to be queried,
on of:
\itemize{
\item SMindex
\item bucket1
\item bucket2
\item deepD
\item runoff
\item totalbucket
Defaults to \dQuote{totalbucket}.
}}

\item{day}{A single day's date to query, \emph{e.g.}, \code{day = "2017-12-31"}, both
\code{Character} and \code{Date} classes are accepted.}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browswer window will open at the
\acronym{TERN} website for you to request a key.}

\item{file}{The filename including the full file path. Defaults to the
current working directory through \code{getwd()}.}

\item{quiet}{Show file download progress bar if TRUE}
}
\value{
A \code{nert.files} object, , a \code{list} of Geotiff files that are locally
available.
}
\description{
Download an entire \acronym{COG} file to your local machine
}
\examples{
download_gtiff()
}
\references{
\url{https://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
\seealso{
Other COGs: 
\code{\link{read_cog}()},
\code{\link{read_cog_dt}()}
}
\author{
Adam Sparks, adam.sparks@curtin.edu.au
}
\concept{COGs}
