% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet.R
\name{extract_aet}
\alias{extract_aet}
\title{Extract AET Values at Point Locations}
\usage{
extract_aet(
  xy,
  month,
  collection = "ETa",
  api_key = get_key(),
  max_tries = 3L,
  initial_delay = 1L
)
}
\arguments{
\item{xy}{Point locations in one of three formats:
\itemize{
\item A named list, \emph{e.g.,}
\code{list("Corrigin" = c(x = 117.87, y = -32.33))}
\item A \code{data.frame} with columns \code{location}, \code{x} (longitude), and \code{y}
(latitude)
\item An \link[sf:sf]{sf::sf} POINT object with a \code{location} column
}}

\item{month}{A month to query, \emph{e.g.,} \code{month = "2023-01-01"} or
\code{month = as.Date("2023-01-01")}.  Both \code{Character} and \code{Date} classes are
accepted.  The value is snapped to the first of the month internally.}

\item{collection}{A \code{character} string of the \acronym{AET} data collection
to be queried:
\itemize{
\item \code{"ETa"}: the primary \acronym{AET} band (mm/month),
\item \code{"pixel_qa"}: the quality assurance flag band.
Defaults to \code{"ETa"}.
}}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browser window will open at the
\acronym{TERN} website for you to request a key.}

\item{max_tries}{An \code{Integer} with the number of times to retry a failed
download before emitting an error message.  Defaults to 3.}

\item{initial_delay}{An \code{Integer} with the number of seconds to delay before
retrying the download.  This increases as the tries increment.  Defaults
to 1.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} with columns \code{location}, \code{x}, \code{y}, and
\verb{aet_\{collection\}_\{YYYYMM\}} (\emph{e.g.,} \code{aet_ETa_202301}).
}
\description{
Extract Actual Evapotranspiration (\acronym{AET}) values from the
\acronym{CMRSET} algorithm data at point locations and return a
\link[data.table:data.table]{data.table::data.table}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

locs <- list(
  "Corrigin" = c(x = 117.87, y = -32.33),
  "Merredin" = c(x = 118.28, y = -31.48),
  "Tamworth" = c(x = 150.84, y = -31.07)
)
tab <- extract_aet(xy = locs, month = "2023-01-01")
tab
\dontshow{\}) # examplesIf}
}
\references{
\acronym{TERN} portal:
\url{https://portal.tern.org.au/metadata/TERN/9fefa68b-dbed-4c20-88db-a9429fb4ba97}

\acronym{DOI}: \url{https://dx.doi.org/10.25901/gg27-ck96}
}
\seealso{
Other COGs: 
\code{\link{read_aet}()},
\code{\link{read_asc}()},
\code{\link{read_smips}()}
}
\concept{COGs}
