% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smips.R
\name{get_smips}
\alias{get_smips}
\title{Get Soil Moisture Integration and Prediction System (SMIPS) Data}
\usage{
get_smips(collection = "totalbucket", day, api_key = get_key())
}
\arguments{
\item{collection}{A character vector of the data collection to be queried,
on of:
\itemize{
\item SMindex
\item bucket1
\item bucket2
\item deepD
\item runoff
\item totalbucket
Defaults to \dQuote{totalbucket}.
}}

\item{day}{A single day's date to query, \emph{e.g.}, \code{day = "2017-12-31"}, both
\code{Character} and \code{Date} classes are accepted.}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browswer window will open at the
\acronym{TERN} website for you to request a key.}
}
\value{
A \link[terra:rast]{terra::rast} object
}
\description{
Soil Moisture Integration and Prediction System (\acronym{SMIPS}) v1.0.
}
\examples{
r <- get_smips(day = "2024-01-01")

# terra::plot() is re-exported for convenience
plot(r)

}
\references{
\url{https://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
