% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain.R
\docType{data}
\name{grain}
\alias{grain}
\title{Synthetic dataset for a grain production experiment in South Australia}
\format{
A \code{data.frame} with 2880 rows and 10 variables.
}
\usage{
data(grain)
}
\description{
Data from a fabricated experiment investigating the effects of nitrogen
application and seeding rate on wheat grain production for a crop conducted
across 10 fictional sites in South Australia.
}
\details{
The variables are as follows:

\describe{
\item{Site}{The experiment site locations. (Factor)}
\item{Latitude}{Latitude for the site. (Numeric)}
\item{Longitude}{Longitude for the site. (Numeric)}
\item{SowDate}{The date that the crops were sown. (Date)}
\item{NitrogenDate}{The date that the nitrogen applications were made.
(Date)}
\item{Rep}{Replicate number. (Each site has 3 replicates.) (Factor)}
\item{Variety}{The anonymised variety (8 levels). (Factor)}
\item{Nitrogen_kgNha}{The applied nitrogen rate treatment, in kg N/ha.
(Factor)}
\item{SeedRate_plantsm2}{The seeding rate, measured in target
establishment plants/m2. (Factor)}
\item{Yield_Tha}{The measured grain yield for the plot in T/ha. (Numeric)}
}
}
\keyword{data}
