% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asc.R
\name{read_asc}
\alias{read_asc}
\title{Read Australian Soil Classification (ASC) Data from TERN}
\usage{
read_asc(
  collection = "EV",
  api_key = get_key(),
  max_tries = NULL,
  initial_delay = NULL
)
}
\arguments{
\item{collection}{The ASC dataset collection variant (default \code{"EV"}). Options:
\describe{
\item{\code{"EV"}}{\strong{(Estimated) Soil Order Class} (character). The
estimated ASC soil order class. Each pixel of the raster maps to one
of 14 soil order classes: Vertosol, Sodosol, Dermosol, Chromosol,
Ferrosol, Kurosol, Tenosol, Kandosol, Hydrosol, Podosol, Rudosol,
Calcarasol, Organosol, Anthroposol.}
\item{\code{"CI"}}{\strong{Confusion Index} (0-1). A unitless index between
0.0 and 1.0, approximating the uncertainty of the Random Forest
classification for the soil order at each pixel. (A higher value of the
confusion index means that the classification at that pixel is more
uncertain.)}
}}

\item{api_key}{A \code{character} string containing your \acronym{TERN}
\acronym{API} key. See \code{\link[=get_key]{get_key()}} for setup.}

\item{max_tries}{Maximum number of download retries before an error is
raised. Default=\code{NULL}, in which case the maximum retry number is
resolved from the option \code{nert.max_tries} if that option exists.
(Defaults to 3 retries if \code{nert.max_tries} has not been set.)}

\item{initial_delay}{Initial retry delay in seconds (doubles with each
attempt). Default=\code{NULL}, in which case the initial delay is
resolved from the option \code{nert.initial_delay} if that option exists.
(Defaults to a 1 second initial delay if \code{nert.initial_delay} has
not been set.)}
}
\value{
A \link[terra:SpatRaster]{terra::SpatRaster} object containing the soil order classes (or the
values of the confusion index).
}
\description{
Wrapper around \code{\link[=read_tern]{read_tern()}} for retrieving Australian Soil Classification
(\acronym{ASC}) soil order classes from the TERN Data Portal. The ASC dataset
provides a comprehensive map at 90m X 90m spatial resolution across
Australia, mapping each spatial pixel to one of 14 soil order classes using
a Random Forest classifier. The dataset also provides a confusion index
raster that estimates the uncertainty (between 0.0 and 1.0) associated with
the classification.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Australian Soil Classification (soil orders as character)
r_asc <- read_asc()
autoplot(r_asc)

# Confusion Index (uncertainty of the classification, higher=more uncertain)
r_ci <- read_asc(collection = "CI")
autoplot(r_ci)
\dontshow{\}) # examplesIf}
}
\references{
Searle, R. (2021). Australian Soil Classification Map. Version 1.
Terrestrial Ecosystem Research Network. (Dataset).
\doi{10.25901/edyr-wg85}.

TERN ASC Point-of-truth metadata URL:
\url{https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/15728dba-b49c-4da5-9073-13d8abe67d7c}
}
\seealso{
\code{\link[=read_tern]{read_tern()}}
}
