% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cog_dt.R
\name{read_cog_dt}
\alias{read_cog_dt}
\title{Read COGs from TERN}
\usage{
read_cog_dt(
  data = "smips",
  collection = "totalbucket",
  day,
  api_key = get_key()
)
}
\arguments{
\item{data}{A character vector of the data source to be queried, currently
only \dQuote{smips}}

\item{collection}{A character vector of the data collection to be queried,
currenly only \dQuote{smips} is supported with the following collections:
\itemize{
\item SMindex
\item bucket1
\item bucket2
\item deepD
\item runoff
\item totalbucket
Defaults to \dQuote{totalbucket}.
}}

\item{day}{A single day's date to query, \emph{e.g.}, \code{day = "2017-12-31"}, both
\code{Character} and \code{Date} classes are accepted.}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browser window will open at the
\acronym{TERN} website for you to request a key.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Read Cloud Optimised Geotiff (\acronym(COG)) files from \acronym{TERN} in
your active \R session as a \CRANpkg{data.table} object.
}
\note{
Currently only Soil Moisture Integration and Prediction System
(\acronym{SMIPS}) v1.0 is supported.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

r <- read_cog_dt(day = "2024-01-01")

r
\dontshow{\}) # examplesIf}
}
\references{
\url{https://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
\seealso{
Other COGs: 
\code{\link{download_gtiff}()},
\code{\link{read_cog}()}
}
\concept{COGs}
