% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smips.R
\name{.check_collection_agreement}
\alias{.check_collection_agreement}
\title{Validate Days Requested Align With Collection}
\usage{
.check_collection_agreement(.collection, .day)
}
\arguments{
\item{.collection}{The user-supplied SMIPS collection being asked for.}

\item{.day}{The user-supplied date being asked for.}
}
\description{
Not all dates are offered by all collections. This checks the user inputs to
be sure that unavailable dates are not requested from collections that do not
provide them.
}
\keyword{internal}
