% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\name{.make_aet_url}
\alias{.make_aet_url}
\title{Build a GDAL vsicurl URL for an AET Collection}
\usage{
.make_aet_url(.collection, .month, .api_key)
}
\arguments{
\item{.collection}{The user-supplied \acronym{AET} collection
(\code{"ETa"} or \code{"pixel_qa"}).}

\item{.month}{The validated \code{POSIXct} date snapped to the first of
the month.}

\item{.api_key}{The \acronym{URL}-encoded \acronym{API} key.}
}
\value{
A \code{character} GDAL vsicurl URL string.
}
\description{
Build a GDAL vsicurl URL for an AET Collection
}
\keyword{internal}
