% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet.R
\name{.make_aet_url}
\alias{.make_aet_url}
\title{Create an AET URL}
\usage{
.make_aet_url(.collection, .month, .api_key)
}
\arguments{
\item{.collection}{The user-supplied \acronym{AET} collection being asked
for (\code{"ETa"} or \code{"pixel_qa"}).}

\item{.month}{The validated \code{POSIXct} date snapped to the first of the
month.}

\item{.api_key}{The \acronym{URL}-encoded \acronym{API} key.}
}
\description{
Builds the full \acronym{GDAL} vsicurl URL for a given \acronym{AET}
collection and month.
}
\keyword{internal}
