% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smips.R
\name{.make_smips_url}
\alias{.make_smips_url}
\title{Create an SMIPS URL}
\usage{
.make_smips_url(.collection, .day)
}
\arguments{
\item{.collection}{The user-supplied SMIPS collection being asked for.}

\item{.day}{The user-supplied date being asked for.}
}
\description{
Creates the SMIPS specific portion of a URL to read or fetch a COG.
}
\keyword{internal}
