% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\docType{data}
\name{.PHENOLOGY_METRICS}
\alias{.PHENOLOGY_METRICS}
\title{Phenology metric metadata: index and directory name}
\format{
An object of class \code{list} of length 2.
}
\usage{
.PHENOLOGY_METRICS
}
\description{
Phenology metric metadata: index and directory name
}
\keyword{datasets}
\keyword{internal}
