% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.read_cog}
\alias{.read_cog}
\title{Read a COG from TERN}
\usage{
.read_cog(full_url, max_tries, initial_delay)
}
\arguments{
\item{full_url}{The URL providing access to the requested data.}

\item{max_tries}{The number of times to retry downloading before timing out.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object of the requested data.
}
\description{
Read a COG from TERN
}
\keyword{internal}
