% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aet.R
\name{.read_tern_aet}
\alias{.read_tern_aet}
\title{Internal handler for retrieving the AET data}
\usage{
.read_tern_aet(dots, api_key, max_tries, initial_delay)
}
\arguments{
\item{dots}{Named list of \code{...} args from \code{\link[=read_tern]{read_tern()}}.}

\item{api_key}{URL-encoded API key.}

\item{max_tries, initial_delay}{Passed to \code{\link[=.read_cog]{.read_cog()}}.}
}
\description{
Internal handler for retrieving the AET data
}
\keyword{internal}
