% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\name{.read_tern_canopy_height}
\alias{.read_tern_canopy_height}
\title{Internal handler for Canopy Height (\code{TERN/36c98155})}
\usage{
.read_tern_canopy_height(api_key, max_tries, initial_delay)
}
\arguments{
\item{api_key}{URL-encoded API key.}

\item{max_tries, initial_delay}{Passed to \code{\link[=.read_cog]{.read_cog()}}.}
}
\description{
Internal handler for Canopy Height (\code{TERN/36c98155})
}
