% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\name{.read_tern_phenology}
\alias{.read_tern_phenology}
\title{Internal handler for Land Surface Phenology (TERN/2bb0c81a)}
\usage{
.read_tern_phenology(dots, api_key, max_tries, initial_delay)
}
\arguments{
\item{dots}{Named list including \code{metric}, \code{year}, \code{season} if provided.}

\item{api_key, max_tries, initial_delay}{Passed to \code{\link[=.read_cog]{.read_cog()}}.}
}
\description{
Reads MODIS-derived Start or End of Growing Season for 2003-2018.
}
\keyword{internal}
\keyword{internal}
