% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tern.R
\docType{data}
\name{.tern_aliases}
\alias{.tern_aliases}
\title{Alias mapping for short dataset names
Maps user-friendly short names (e.g. "SMIPS", "AWC") to dispatch IDs}
\format{
An object of class \code{list} of length 14.
}
\usage{
.tern_aliases
}
\description{
Alias mapping for short dataset names
Maps user-friendly short names (e.g. "SMIPS", "AWC") to dispatch IDs
}
\keyword{datasets}
\keyword{internal}
