% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{Get or Set Up API Key for TERN}
\usage{
get_key()
}
\value{
A string value with your \acronym{API} key value.
}
\description{
Checks first to get key from your .Rprofile or .Renviron (or similar) file.
If it's not found, then it suggests setting it up.  Can be used to check
that your key that \R is using is the key that you wish to be using or for
guidance in setting up the keys.
}
\section{Requesting an API Key}{
To request an \acronym{API} key, go to
\url{https://account.tern.org.au/authenticated_user/apikeys} and click on
"Sign In" in the upper right corner. Sign in with your proper credentials.
Then, from the left-hand menu, click on "Create API Key".  Once this is
done, copy the key and put it in your .Renviron using
\code{\link[usethis:edit_r_environ]{usethis::edit_r_environ()}} as \code{TERN_API_KEY="your_api_key"}. Restart your
\R session and the query should work.
}

\examples{
\dontrun{
get_key()
}

}
