% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{Get or Set Up API Key for TERN}
\usage{
get_key()
}
\value{
A string value with your \acronym{API} key value.
}
\description{
Checks first to get key from your .Rprofile or .Renviron (or similar) file.
If it's not found, then it suggests setting it up.  Can be used to check
that your key that \R is using is the key that you wish to be using or for
guidance in setting up the keys.
}
\details{
The suggestion is to use your .Renviron to set up the API key. However, if
you regularly interact with the APIs outside of \R using some other
language you may wish to set these up in your .bashrc, .zshrc, or
config.fish for cross-language use.
}
\note{
\acronym{TERN} creates \acronym{API} keys that have special characters that
include \dQuote{/}, which causes the query to fail. Currently, \code{get_key()}
checks for this in the \code{API_KEY} string and replaces it with \dQuote(\%2f)
so that the query will work properly.
}
\examples{
\dontrun{
get_key()
}

}
