% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smips.R
\name{get_smips}
\alias{get_smips}
\title{Get Soil Moisture Integration and Prediction System (SMIPS) Data}
\usage{
get_smips(start_date, end_date)
}
\description{
Soil Moisture Integration and Prediction System (\acronym{SMIPS})
}
\references{
\url{https://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
