% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nert-package.R
\docType{package}
\name{nert-package}
\alias{nert}
\alias{nert-package}
\title{nert: Curated Access to TERN Environmental Raster Data}
\description{
Provides curated access to gridded environmental data published by Australia's Terrestrial Ecosystem Research Network (TERN; \url{https://www.tern.org.au/}), spanning national soil attribute, soil classification, soil biodiversity, soil moisture, actual evapotranspiration, vegetation canopy height and phenology products. The package reads Cloud-Optimised GeoTIFF and virtual raster files directly from the TERN data server via GDAL '/vsicurl' range requests, streaming only the requested spatial and temporal subset for efficient data transfer. The environmental raster data are returned as terra::SpatRaster objects. A convenient helper function is also provided that provides a batch-download capability given specific locations and dates, returning the extracted data values in an analysis-ready data.table. Access to the TERN datasets using this package requires a TERN API key.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagi-aus.github.io/nert/}
  \item \url{https://github.com/AAGI-AUS/nert}
  \item Report bugs at \url{https://github.com/AAGI-AUS/nert/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{sam.rogers@adelaide.edu.au} (\href{https://orcid.org/0000-0002-8147-1239}{ORCID})

Authors:
\itemize{
  \item Sam Rogers \email{sam.rogers@adelaide.edu.au} (\href{https://orcid.org/0000-0002-8147-1239}{ORCID})
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
  \item Wasin Pipattungsakul \email{wasin.pipattungsakul@adelaide.edu.au}
  \item Russell Edson \email{russell.edson@adelaide.edu.au} (\href{https://orcid.org/0000-0002-4607-5396}{ORCID})
  \item Max Moldovan \email{max.moldovan@adelaide.edu.au} (\href{https://orcid.org/0000-0001-9680-8474}{ORCID})
}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (\href{https://ror.org/02xwr1996}{ROR}) (GRDC Project CUR2210-005OPX (AAGI-CU), UOA2301-005OPX (AAGI-AU)) [funder, copyright holder]
}

}
\keyword{internal}
