% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nert-package.R
\docType{package}
\name{nert-package}
\alias{nert}
\alias{nert-package}
\title{nert: An API Client for TERN Data}
\description{
Provides access to Australia's TERN (Terrestrial Ecosystem Research Network) data through the API, \url{https://tern.org.au}.
}
\keyword{internal}
