% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nert-package.R
\docType{package}
\name{nert-package}
\alias{nert}
\alias{nert-package}
\title{nert: An API Client for TERN Data}
\description{
Provides access to Australia's TERN (Terrestrial Ecosystem Research Network) data through the API, \url{https://tern.org.au}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagi-aus.github.io/nert/}
  \item \url{https://github.com/AAGI-AUS/nert}
  \item Report bugs at \url{https://github.com/AAGI-AUS/nert/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{sam.rogers@adelaide.edu.au} (\href{https://orcid.org/0000-0002-8147-1239}{ORCID})

Authors:
\itemize{
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
  \item Wasin Pipattungsakul \email{wasin.pipattungsakul@adelaide.edu.au}
  \item Russell Edson \email{russell.edson@adelaide.edu.au} (\href{https://orcid.org/0000-0002-4607-5396}{ORCID})
  \item Max Moldovan \email{max.moldovan@adelaide.edu.au}
}

Other contributors:
\itemize{
  \item Curtin University (\href{https://ror.org/02n415q13}{ROR}) (Provided support through Adam Sparks's time.) [copyright holder]
  \item Grains Research and Development Corporation (\href{https://ror.org/02xwr1996}{ROR}) (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
