% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nert-package.R
\docType{package}
\name{nert-package}
\alias{nert}
\alias{nert-package}
\title{nert: An API Client for TERN Data}
\description{
Provides access to Australia's TERN (Terrestrial Ecosystem Research Network) data through the API, \url{https://tern.org.au}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagi-aus.github.io/nert/}
  \item \url{https://github.com/AAGI-AUS/nert}
  \item Report bugs at \url{https://github.com/AAGI-AUS/nert/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Wasin Pipattungsakul \email{wasin.pipattungsakul@adelaide.edu.au}
  \item Russell Edson \email{russell.edson@adelaide.edu.au} (\href{https://orcid.org/0000-0002-4607-5396}{ORCID})
  \item Sam Rogers \email{sam.rogers@adelaide.edu.au} (\href{https://orcid.org/0000-0002-4607-5396}{ORCID})
  \item Max Moldovan \email{max.moldovan@adelaide.edu.au}
}

}
\keyword{internal}
