% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet.R
\name{read_aet}
\alias{read_aet}
\title{Read AET COGs from TERN}
\usage{
read_aet(
  month,
  collection = "ETa",
  api_key = get_key(),
  max_tries = 3L,
  initial_delay = 1L
)
}
\arguments{
\item{month}{A month to query, \emph{e.g.,} \code{month = "2023-01-01"} or
\code{month = as.Date("2023-01-01")}.  Both \code{Character} and \code{Date} classes are
accepted.  The value is snapped to the first of the month internally.}

\item{collection}{A \code{character} string of the \acronym{AET} data collection
to be queried:
\itemize{
\item \code{"ETa"}: the primary \acronym{AET} band (mm/month),
\item \code{"pixel_qa"}: the quality assurance flag band.
Defaults to \code{"ETa"}.
}}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browser window will open at the
\acronym{TERN} website for you to request a key.}

\item{max_tries}{An \code{Integer} with the number of times to retry a failed
download before emitting an error message.  Defaults to 3.}

\item{initial_delay}{An \code{Integer} with the number of seconds to delay before
retrying the download.  This increases as the tries increment.  Defaults
to 1.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object of the national mosaic for the requested
month.
}
\description{
Read Actual Evapotranspiration (\acronym{AET}) using the \acronym{CMRSET}
algorithm Cloud Optimised Geotiff (\acronym{COG}) files from \acronym{TERN}
in your \R session.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

r <- read_aet(month = "2023-01-01")

# `tidyterra::autoplot` is re-exported for convenience
autoplot(r)
\dontshow{\}) # examplesIf}
}
\references{
\acronym{TERN} portal:
\url{https://portal.tern.org.au/metadata/TERN/9fefa68b-dbed-4c20-88db-a9429fb4ba97}

\acronym{DOI}: \url{https://dx.doi.org/10.25901/gg27-ck96}
}
\seealso{
Other COGs: 
\code{\link{extract_aet}()},
\code{\link{read_asc}()},
\code{\link{read_smips}()}
}
\concept{COGs}
