% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cog.R
\name{read_cog}
\alias{read_cog}
\title{Read COGs from TERN}
\usage{
read_cog(
  dataset = "smips",
  collection = "totalbucket",
  day,
  api_key = get_key(),
  max_tries = 3,
  initial_delay = 1
)
}
\arguments{
\item{dataset}{A character vector of the data source to be queried, currently
only \dQuote{SMIPS}.}

\item{collection}{A character vector of the data collection to be queried,
currently only \dQuote{SMIPS} is supported with the following collections:
\itemize{
\item SMindex
\item bucket1
\item bucket2
\item deepD
\item runoff
\item totalbucket
Defaults to \dQuote{totalbucket}. Multiple \code{collections} are supported,
\emph{e.g.}, \code{collection = c("SMindex", "totalbucket")}.
}}

\item{day}{A vector of date(s) to query, \emph{e.g.}, \code{day = "2017-12-31"} or
\code{day = seq.Date(as.Date("2017-12-01"), as.Date("2017-12-31"), "days")}, both
\code{Character} and \code{Date} classes are accepted.}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browser window will open at the
\acronym{TERN} website for you to request a key.}

\item{max_tries}{An integer \code{Integer} with the number of times to retry a
failed download before emitting an error message.  Defaults to 3.}

\item{initial_delay}{An \code{Integer} with the number of seconds to delay before
retrying the download.  This increases as the tries increment.  Defaults to
1.}
}
\value{
A \link[terra:rast]{terra::rast} object
}
\description{
Read Cloud Optimised Geotiff (\acronym{COG}) files from \acronym{TERN} in
your active \R session.
}
\note{
Currently only Soil Moisture Integration and Prediction System
(\acronym{SMIPS}) v1.0 is supported.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

r <- read_cog(day = "2024-01-01")

# terra::plot() is re-exported for convenience
plot(r)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
\concept{COGs}
