% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenology.R
\name{read_phenology}
\alias{read_phenology}
\title{Read Land Surface Phenology from TERN}
\usage{
read_phenology(
  year = NULL,
  season = 1L,
  collection = "SGS",
  api_key = get_key(),
  max_tries = NULL,
  initial_delay = NULL
)
}
\arguments{
\item{year}{An integer year (2003--2018).}

\item{season}{Season number: \code{1} (default) or \code{2}.}

\item{collection}{Phenology metric abbreviation.  One of \code{"SGS"}
(default), \code{"PGS"}, \code{"EGS"}, \code{"LGS"}, \code{"EVI1"},
\code{"EVI2"}, \code{"EVIP"}, or \code{"EVII"}.}

\item{api_key}{A \code{character} string containing your \acronym{TERN}
\acronym{API} key. Defaults to automatic detection from your
\code{.Renviron} or \code{.Rprofile}.  See \code{\link[=get_key]{get_key()}} for setup.}

\item{max_tries}{Maximum number of download retries before an error is
raised. Default=\code{NULL}, in which case the maximum retry number is
resolved from the option \code{nert.max_tries} if that option exists.
(Defaults to 3 retries if \code{nert.max_tries} has not been set.)}

\item{initial_delay}{Initial retry delay in seconds (doubles with each
attempt). Default=\code{NULL}, in which case the initial delay is
resolved from the option \code{nert.initial_delay} if that option exists.
(Defaults to a 1 second initial delay if \code{nert.initial_delay} has
not been set.)}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object of the requested phenology metric.
}
\description{
Wrapper around \code{\link[=read_tern]{read_tern()}} for retrieving the Australian land surface
phenology dataset (v1.0) from the TERN Data Portal. This data comprises
phenology metrics from two growing seasons per year between 2003 and
2018 (inclusive), estimated at 500m X 500m spatial resolution across
all of Australia using thresholded MODIS Enhanced Vegetation Index (EVI)
imagery.
}
\section{Phenology metrics}{

Eight phenological metrics are available via the \code{collection}
argument:
\describe{
\item{\code{"SGS"}}{\strong{Start of Growing Season} (default). A number
between -150 and 345 indicating the start of the growing season (with
negative numbers meaning the prior year).}
\item{\code{"PGS"}}{\strong{Peak of Growing Season}. A number between 9 and
361 indicating the day of the year for the peak of the growing season.}
\item{\code{"EGS"}}{\strong{End of Growing Season}. A number between 25 and
519 indicating the day of the year that marks the end of the growing
season, with numbers above 365 (or 366 for leap years) meaning the
following year.}
\item{\code{"LGS"}}{\strong{Length of Growing Season}, measured in days.}
\item{\code{"EVI1"}}{\strong{Minimum of EVI before peak}. A unitless index
between 0 and 10000 (i.e., scaled by 10000) for the minimum
value of the Enhanced Vegetation Index \emph{before} the peak of the growing
season.}
\item{\code{"EVI2"}}{\strong{Minimum of EVI after peak}. A unitless index
between 0 and 10000 (i.e., scaled by 10000) for the minimum
value of the Enhanced Vegetation Index \emph{after} the peak of the growing
season.}
\item{\code{"EVIP"}}{\strong{Peak EVI}. A unitless index between 0 and 10000
(i.e., scaled by 10000) for the value of the Enhanced Vegetation Index
\emph{at} the peak of the growing season.}
\item{\code{"EVII"}}{\strong{Integral of the EVI across the season}. The
calculated integral of the Enhanced Vegetation Index (including the
factor of 10000 scaling) under the growing season cycle curve.}
}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Read Start of Growing Season for 2018, Season 1
r <- read_phenology(year = 2018)
autoplot(r)

# Read End of Growing Season for 2015, Season 2
r_egs <- read_phenology(year = 2015, season = 2, collection = "EGS")
autoplot(r_egs)
\dontshow{\}) # examplesIf}
}
\references{
Xie, Q. & Huete, A. (2024). Australian land surface phenology dataset at
500m resolution. Version 1.0. Terrestrial Ecosystem Research Network.
(Dataset). URL: \url{https://portal.tern.org.au/metadata/2bb0c81a-41a9-434c-b87a-db0301cb52fb}.

TERN Land Surface Phenology Point-of-truth metadata URL:
\url{https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/2bb0c81a-5a09-4a0e-bd86-5cd2be8def26}
}
\seealso{
\code{\link[=read_tern]{read_tern()}}
}
