% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soil_diversity.R
\name{read_soil_diversity}
\alias{read_soil_diversity}
\title{Read Soil Beta Diversity from TERN}
\usage{
read_soil_diversity(
  collection = "Bacteria",
  axis = 1L,
  api_key = get_key(),
  max_tries = NULL,
  initial_delay = NULL
)
}
\arguments{
\item{collection}{One of \code{"Bacteria"} (default) or \code{"Fungi"}.}

\item{axis}{\acronym{NMDS} axis number: \code{1} (default), \code{2}, or
\code{3}.}

\item{api_key}{A \code{character} string containing your \acronym{TERN}
\acronym{API} key. Defaults to automatic detection from your
\code{.Renviron} or \code{.Rprofile}. See \code{\link[=get_key]{get_key()}} for setup.}

\item{max_tries}{Maximum number of download retries before an error is
raised. Default=\code{NULL}, in which case the maximum retry number is
resolved from the option \code{nert.max_tries} if that option exists.
(Defaults to 3 retries if \code{nert.max_tries} has not been set.)}

\item{initial_delay}{Initial retry delay in seconds (doubles with each
attempt). Default=\code{NULL}, in which case the initial delay is
resolved from the option \code{nert.initial_delay} if that option exists.
(Defaults to a 1 second initial delay if \code{nert.initial_delay} has
not been set.)}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object of the requested Beta Diversity NMDS axis.
}
\description{
Wrapper around \code{\link[=read_tern]{read_tern()}} for retrieving the v1.0 datasets for
soil bacteria and soil fungi Beta Diversity from the TERN Data Portal.
The Beta Diversity datasets were constructed using Digital Soil Mapping
techniques together with Biome of Australia Soil Environments (BASE)
DNA sequences, and provide Non-metric MultiDimensional Scaling (NMDS)
axes for soil bacteria and fungi at 90m X 90m spatial resolution across
Australia.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Read bacterial diversity NMDS axis 1
r <- read_soil_diversity()
autoplot(r)

# Read fungal diversity NMDS axis 2
r_f2 <- read_soil_diversity(collection = "Fungi", axis = 2)
autoplot(r_f2)
\dontshow{\}) # examplesIf}
}
\references{
Dobarco, M., Wadoux, A. & Xue, P. (2024). Soil and Landscape Grid National
Soil Attribute Maps - Soil Bacteria and Fungi Beta Diversity (3"
resolution) - Release 1. Version 1.0. Terrestrial Ecosystem Research
Network. (Dataset). \doi{10.25919/4x7n-y874}.

TERN Soil Beta Diversity Point-of-truth metadata URL:
\url{https://geonetwork.tern.org.au/geonetwork/srv/eng/catalog.search#/metadata/4a428d52-dda6-4097-8dd9-d3ec63973029}
}
\seealso{
\code{\link[=read_tern]{read_tern()}}
}
